benchmark/downstream/asr/requirements.txt
tqdm
pandas
future
joblib
pyyaml
sentencepiece
tb-nightly
torch>=1.2.0
torchaudio
matplotlib
librosa
audiomentations
editdistance